Recurse submodules on clone and checkout, use Release build
Submodules will have to be recursively initialized starting with v2.3.x,
so I figured why not just prepare for that. This is due to
https://github.com/BeamMP/BeamMP-Server/pull/44.

Release build will reduce binary size and enables us to get rid of the
`strip` step entirely. The server will also run considerably faster.

Running make with -j $(nproc) should fetch the amount of available cores
and use that, instead of the hardcoded 4. Since BeamMP-Server has quite
a few dependencies, this will be a welcome improvement on machines with
lots of threads.